Moved all the related classes into same packages
